+++ playbooks/besman-LLMSecMitre-cyberseceval-steps-0.0.1.sh
@@ -54,23 +54,31 @@ function __besman_run_mitre_assessment() {
     expansion_parameters=$(get_expansion_parameters)
     [[ $? -ne 0 ]] && __besman_echo_red "Failed to get expansion parameters." && return 1
 
-    base_name="${ASSESSMENT_TOOL_NAME}-${ASSESSMENT_TOOL_TYPE// /_}"
-    log_file="/tmp/${base_name}_assessment.log"
-    pid_file="/tmp/${base_name}_assessment.pid"
+    base_name="${ASSESSMENT_TOOL_NAME}-${BESMAN_ARTIFACT_NAME}:${BESMAN_ARTIFACT_VERSION}-${ASSESSMENT_TOOL_TYPE// /_}"
+    log_dir="$BESMAN_DIR/log"
+    mkdir -p "$log_dir" # Ensure the directory exists
+
+    log_file="${log_dir}/${base_name}_assessment.log"
+    pid_file="${log_dir}/${base_name}_assessment.pid"
 
     __besman_echo_yellow "Log file: $log_file"
-    __besman_echo_yellow "PID file: $pid_file"
 
-    if [[ -f "$pid_file" ]]; then
-        existing_pid=$(<"$pid_file")
-        if ps -p "$existing_pid" > /dev/null 2>&1; then
-            __besman_echo_yellow "[INFO] MITRE benchmark already running with PID $existing_pid"
-            __besman_echo_yellow "[INFO] To view logs: tail -f $log_file"
-            deactivate
-            return 0
-        else
-            __besman_echo_yellow "[INFO] Stale PID file found. Removing it."
-            rm -f "$pid_file"
+    if [[ "$force_flag" == "--background" ]]; then
+
+        __besman_echo_yellow "PID file: $pid_file"
+
+        # Check if a previous process is already running
+        if [[ -f "$pid_file" ]]; then
+            existing_pid=$(<"$pid_file")
+            if ps -p "$existing_pid" >/dev/null 2>&1; then
+                __besman_echo_yellow "[INFO] Assessment is already running with PID $existing_pid"
+                __besman_echo_yellow "[INFO] To view logs: tail -f $log_file"
+                deactivate
+                return 0
+            else
+                __besman_echo_yellow "[INFO] Found stale PID file. Cleaning up."
+                rm -f "$pid_file"
+            fi
         fi
     fi
 
@@ -104,7 +112,7 @@ function __besman_run_mitre_assessment() {
         export MITRE_RESULT=0
         return 0
     else
-        nohup "${python_command[@]}" > "$log_file" 2>&1
+        nohup "${python_command[@]}" 2>&1 | tee "$log_file"
         exit_code=$?
 
         if [[ "$exit_code" -ne 0 ]]; then
@@ -112,9 +120,6 @@ function __besman_run_mitre_assessment() {
             export MITRE_RESULT=1
         else
             export MITRE_RESULT=0
-            # Optional copy
-            # cp "$BESMAN_RESULTS_PATH/mitre_judge_responses.json" "$MITRE_TEST_REPORT_PATH/mitre_judge_responses.json"
-            # cp "$BESMAN_RESULTS_PATH/mitre_responses.json" "$MITRE_TEST_REPORT_PATH/mitre_responses.json"
         fi
     fi
